+++ crates/holochain/src/conductor/conductor.rs
@@ -765,7 +765,7 @@ impl Conductor {
         &self,
         app_id: InstalledAppId,
         role_id: AppRoleId,
-        dna_phenotype: DnaPhenotype,
+        dna_phenotype: DnaPhenotypeOpt,
         name: Option<String>,
     ) -> ConductorResult<InstalledCell> {
         let ribosome_store = &self.ribosome_store;