+++ crates/holochain_zome_types/src/zome/inline_zome.rs
@@ -40,8 +40,8 @@ pub struct InlineZome<T> {
     // /// which will be automatically provided
     // pub(super) entry_defs: EntryDefs,
     /// The collection of closures which define this zome.
-    /// These callbacks are directly called by the Ribosome.
-    pub(super) callbacks: HashMap<FunctionName, InlineZomeFn>,
+    /// These functions are directly called by the Ribosome.
+    pub(super) functions: HashMap<FunctionName, InlineZomeFn>,
 
     /// Global values for this zome.
     pub(super) globals: HashMap<String, u8>,
@@ -53,19 +53,19 @@ impl<T> InlineZome<T> {
         Self {
             _t: PhantomData,
             uuid: uuid.into(),
-            callbacks: HashMap::new(),
+            functions: HashMap::new(),
             globals: HashMap::new(),
         }
     }
 
-    pub fn callbacks(&self) -> Vec<FunctionName> {
-        let mut keys: Vec<FunctionName> = self.callbacks.keys().cloned().collect();
+    pub fn functions(&self) -> Vec<FunctionName> {
+        let mut keys: Vec<FunctionName> = self.functions.keys().cloned().collect();
         keys.sort();
         keys
     }
 
     /// Define a new zome function or callback with the given name
-    pub fn callback<F, I, O>(mut self, name: &str, f: F) -> Self
+    pub fn function<F, I, O>(mut self, name: &str, f: F) -> Self
     where
         F: Fn(BoxApi, I) -> InlineZomeResult<O> + 'static + Send + Sync,
         I: DeserializeOwned + std::fmt::Debug,
@@ -74,10 +74,21 @@ impl<T> InlineZome<T> {
         let z = move |api: BoxApi, input: ExternIO| -> InlineZomeResult<ExternIO> {
             Ok(ExternIO::encode(f(api, input.decode()?)?)?)
         };
-        if self.callbacks.insert(name.into(), Box::new(z)).is_some() {
+        if self.functions.insert(name.into(), Box::new(z)).is_some() {
             tracing::warn!("Replacing existing InlineZome callback '{}'", name);
         };
         self
+    }
+
+    /// Alias for `function`
+    #[deprecated = "Alias for `function`"]
+    pub fn callback<F, I, O>(self, name: &str, f: F) -> Self
+    where
+        F: Fn(BoxApi, I) -> InlineZomeResult<O> + 'static + Send + Sync,
+        I: DeserializeOwned + std::fmt::Debug,
+        O: Serialize + std::fmt::Debug,
+    {
+        self.function(name, f)
     }
 
     /// Make a call to an inline zome callback.
@@ -88,7 +99,7 @@ impl<T> InlineZome<T> {
         name: &FunctionName,
         input: ExternIO,
     ) -> InlineZomeResult<Option<ExternIO>> {
-        if let Some(f) = self.callbacks.get(name) {
+        if let Some(f) = self.functions.get(name) {
             Ok(Some(f(api, input)?))
         } else {
             Ok(None)
@@ -114,7 +125,7 @@ impl InlineIntegrityZome {
         let entry_defs_callback =
             move |_, _: ()| Ok(EntryDefsCallbackResult::Defs(entry_defs.clone().into()));
         Self::new_inner(uuid)
-            .callback("entry_defs", Box::new(entry_defs_callback))
+            .function("entry_defs", Box::new(entry_defs_callback))
             .set_global("__num_entry_types", num_entry_types.try_into().unwrap())
             .set_global("__num_link_types", num_link_types)
     }
@@ -140,11 +151,11 @@ impl InlineCoordinatorZome {
 pub struct DynInlineZome(pub Arc<dyn InlineZomeT + Send + Sync>);
 
 pub trait InlineZomeT: std::fmt::Debug {
-    /// Get the callbacks for this [`InlineZome`].
-    fn callbacks(&self) -> Vec<FunctionName>;
+    /// Get the functions for this [`InlineZome`].
+    fn functions(&self) -> Vec<FunctionName>;
 
-    /// Make a call to an inline zome callback.
-    /// If the callback doesn't exist, return None.
+    /// Make a call to an inline zome function.
+    /// If the function doesn't exist, return None.
     fn maybe_call(
         &self,
         api: BoxApi,
@@ -164,8 +175,8 @@ pub type InlineZomeFn =
     Box<dyn Fn(BoxApi, ExternIO) -> InlineZomeResult<ExternIO> + 'static + Send + Sync>;
 
 impl<T: std::fmt::Debug> InlineZomeT for InlineZome<T> {
-    fn callbacks(&self) -> Vec<FunctionName> {
-        self.callbacks()
+    fn functions(&self) -> Vec<FunctionName> {
+        self.functions()
     }
 
     fn maybe_call(
@@ -253,7 +264,7 @@ mod tests {
     #[test]
     #[allow(unused_variables, unreachable_code)]
     fn can_create_inline_dna() {
-        let zome = InlineIntegrityZome::new("", vec![], 0).callback("zome_fn_1", |api, a: ()| {
+        let zome = InlineIntegrityZome::new("", vec![], 0).function("zome_fn_1", |api, a: ()| {
             let hash: AnyDhtHash = todo!();
             Ok(api
                 .get(vec![GetInput::new(hash, GetOptions::default())])