+++ crates/hdk/CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## 0.0.153
+
+## 0.0.152
+
 ## 0.0.151
 
 ## 0.0.150