+++ crates/holochain_zome_types/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## \[Unreleased\]
+
+## 0.0.49
+
+## 0.0.48
 
 - Add function to set DNA name. [\#1547](https://github.com/holochain/holochain/pull/1547)